Add files via upload
We have created a Mini ATM SYSTEM in which intially a person’s balance is zero. Here the account is considered as pre defined. He/She need to deposit some amount to take further actions . He/She can able to withdraw amount only more than 100 since which is the general rule in an ATM in our day to day life.